Foresee a next state

Get a predictive next position for a pedestrian (R function is used in testing. Another copy of C++ function runs the model fitting.)

Usage

foresee_next(x0, v, pt, t0 = 0)

Arguments

x0

an initial (pedestrian) position

v

pedestrian walking speed

pt

how far in time a pedestrian can foresee what consequence its action may result

t0

how much time is needed to enact an action; assuming instantaneous when t0 = 0; this should map to t0.

Details

Function to get the predictive pedestrian state if beginning to cross. I assume a pedestrian prepares to walk to the north on the negative side of y axis.

Examples

## some examples