Crossing decisions data

ped stores three data frames, d, d0 and dat. snew and sold are two labelling schemes for participant labels. The multiple data frames are to enable different software to process conveniently. d0 has all the variables being aggregated across participants. dat reduces the columns. d is to be processed by RStan via rethinking package.

Trial. A label recording also the sequence of trial appearing in the task.

Jitter. A randomly selected time between -0.1 to 0.1 s. This added a small variability to the time-to##' -arrival so as to prevent responses based on anticipating the imperative stimulus.

BeforeCarPassed. A Boolean variable, indicating a response is made before the car passes the crossing zone.

E. A nominal variable, indicating whether a trial is from the behavioural study, exp, or from the EEG study, eeg.

TTA. Time-to-arrival in second. 2.5, 3, 3.5 or 4 seconds.

Side. A nominal variable, indicating a trial showing the avator, at the beginning, stands on the right or left side of scene.

R. A nominal variable, indicating a trial resulting in a safe response or a collision.

C. A Boolean variable, indicating a trial resulting in a safe response (TRUE) or a collision (FALSE).

G. A nominal variable, indicating a response is made before the car passes the crossing zone.

TTANme. A nominal variable for the time-to-arrival.

RT. A continuous variable, indicating response times in second.

s. A nominal variable, indicating different participants.

D. A continuous variable of vehicle initial distance. This column has only one value, i.e., 40 m. This is to calculate vehicle kinematics.

DTTA. A nominal label of distance and TTA for the purpose of plotting figures.

Usage

data(ped)

Details

Behavioural data from human participants doing a road-crossing task.

References

Institute for Transport Studies, University of Leeds. This work was supported by Engineering and Physical Sciences Research Council (EPSRC) under Grant number, EP/S005056/1.

Author

Yi-Shin Lin <yishinlin@pm.me>

Examples

data(ped)
dplt <- dat[RT <= 6 & RT > .1]
1 - nrow(dplt) / nrow(dat)
#> [1] 0.002103365

# Simple get_break function
rt <- dplt$RT
binsize <- psych::describe(rt)$se  
bk <- seq(min(rt)-binsize, max(rt)+binsize, by = binsize)

dv <- data.frame(TTA = c(2.5, 3, 3.5, 4, 2.5, 3, 3.5, 4),
                 Etta = c("exp, 2.5", "exp, 3", "exp, 3.5", "exp, 4",
                          "eeg, 2.5", "eeg, 3", "eeg, 3.5", "eeg, 4"))

# sort(unique(DT$TTA))
# [1] 2.5 3.0 3.5 4.0 

dplt$Rgp <- factor( paste0(dplt$G, ", ", dplt$R), 
                    levels = c("after, safe", "after, hit", "before, safe", "before, hit"))
table(dplt$Rgp)
#> 
#>  after, safe   after, hit before, safe  before, hit 
#>         4269            7         4577         1110 
if (FALSE) { 
p0 <- ggplot() +
  geom_histogram(data = dplt[E=="exp"], aes(x = RT, fill = Rgp), 
                 breaks = bk) +
  geom_histogram(data = dplt[E=="eeg"], aes(x = RT, fill = Rgp), 
                 breaks = bk) +
  ## scale_fill_manual(values = Manu::get_pal("Tui")) +
  geom_vline(data = dv, aes(xintercept = TTA), linetype = "dashed") +
  scale_x_continuous(name = "RT (s)", breaks = c(0, 1, 2, 3, 4, 5, 6)) +
  facet_grid(E~., switch = "y") +
  theme_minimal(base_size = 16) +
  theme(legend.position = "none",
        # theme(legend.position = c(.80, .90),
        legend.title = element_blank(),
        strip.placement = "outside")

dplt$Etta <- factor( paste0(dplt$E, ", ", dplt$TTA), 
                     levels = c("exp, 2.5", "exp, 3", "exp, 3.5", "exp, 4",
                                "eeg, 2.5", "eeg, 3", "eeg, 3.5", "eeg, 4"))

p1 <- ggplot() +
  geom_histogram(data = dplt[E=="exp"], aes(x = RT, fill = Rgp), 
                 breaks = bk) +
  geom_histogram(data = dplt[E=="eeg"], aes(x = RT, fill = Rgp), 
                 breaks = bk) +
  ## scale_fill_manual(values = Manu::get_pal("Tui")) +
  geom_vline(data = dv, aes(xintercept = TTA), linetype = "dashed") +
  scale_x_continuous(name = "RT (s)", breaks = c(0, 1, 2, 3, 4, 5, 6)) +
  ylab("") +
  facet_grid(Etta~., switch = "y") +
  theme_minimal(base_size = 16) +
  theme(legend.position = c(.85, .95),
        legend.title = element_blank(),
        strip.placement = "outside")

gridExtra::grid.arrange(p0, p1, ncol = 2)
}